feat(ExerciseCard): Allow adding/removing sets
Introduces functionality to dynamically add or remove sets for an exercise within the log.

When adding sets, the weight is automatically inherited from the last recorded set, or the base weight if no sets exist. When removing sets, they are removed from the end of the list. This enhances user flexibility in tracking workout progression.